+++ client/src/lib/hooks/useConnection.ts
@@ -1,4 +1,3 @@
-import { Client } from "@modelcontextprotocol/sdk/client/index.js";
 import {
   SSEClientTransport,
   SseError,
@@ -15,8 +14,6 @@ import {
   ListRootsRequestSchema,
   ResourceUpdatedNotificationSchema,
   LoggingMessageNotificationSchema,
-  Request,
-  Result,
   ServerCapabilities,
   PromptReference,
   ResourceReference,
@@ -37,7 +34,6 @@ import { ConnectionStatus } from "../constants";
 import { Notification, StdErrNotificationSchema } from "../notificationTypes";
 import { auth } from "@modelcontextprotocol/sdk/client/auth.js";
 import { InspectorOAuthClientProvider } from "../auth";
-import packageJson from "../../../package.json";
 import {
   getMCPProxyAddress,
   getMCPServerRequestMaxTotalTimeout,
@@ -45,14 +41,9 @@ import {
 } from "@/utils/configUtils";
 import { getMCPServerRequestTimeout } from "@/utils/configUtils";
 import { InspectorConfig } from "../configurationTypes";
-import { Anthropic } from "@anthropic-ai/sdk";
-import {
-  MessageParam,
-  Tool,
-} from "@anthropic-ai/sdk/resources/messages/messages.mjs";
-import readline from "readline/promises";
+import { MCPClient, ExtendedMcpClient } from "../mcpClient";
 
-interface UseConnectionOptions {
+interface McpServerConnection {
   transportType: "stdio" | "sse" | "streamable-http";
   command: string;
   args: string;
@@ -68,15 +59,10 @@ interface UseConnectionOptions {
   onPendingRequest?: (request: any, resolve: any, reject: any) => void;
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   getRoots?: () => any[];
+  client?: MCPClient
 }
 
-// Add interface for extended MCP client with Anthropic
-interface ExtendedMcpClient extends Client {
-  anthropic: Anthropic;
-  processQuery: (query: string, tools: Tool[]) => Promise<string>;
-  chatLoop: (tools: Tool[]) => Promise<void>;
-  cleanup: () => Promise<void>;
-}
+
 
 export function useConnection({
   transportType,
@@ -92,13 +78,14 @@ export function useConnection({
   onStdErrNotification,
   onPendingRequest,
   getRoots,
-}: UseConnectionOptions) {
+  client,
+}: McpServerConnection) {
   const [connectionStatus, setConnectionStatus] =
     useState<ConnectionStatus>("disconnected");
   const { toast } = useToast();
   const [serverCapabilities, setServerCapabilities] =
     useState<ServerCapabilities | null>(null);
-  const [mcpClient, setMcpClient] = useState<ExtendedMcpClient | null>(null);
+  const [mcpClient, setMcpClient] = useState<ExtendedMcpClient | undefined>(client);
   const [requestHistory, setRequestHistory] = useState<
     { request: string; response?: string }[]
   >([]);
@@ -107,9 +94,12 @@ export function useConnection({
   // Add a method to update the API key on the existing client
   const updateApiKey = (newApiKey: string) => {
     if (mcpClient && mcpClient.anthropic) {
-      mcpClient.anthropic = new Anthropic({
-        apiKey: newApiKey,
-        dangerouslyAllowBrowser: true,
+      // Import Anthropic dynamically to avoid import issues
+      import("@anthropic-ai/sdk").then(({ Anthropic }) => {
+        mcpClient.anthropic = new Anthropic({
+          apiKey: newApiKey,
+          dangerouslyAllowBrowser: true,
+        });
       });
     }
   };
@@ -294,252 +284,14 @@ export function useConnection({
     return false;
   };
 
-  const connect = async (_e?: unknown, retryCount: number = 0) => {
-    class MCPClient extends Client<Request, Notification, Result> {
-      anthropic: Anthropic;
-      constructor() {
-        super(
-          {
-            name: "mcpjam-inspector",
-            version: packageJson.version,
-          },
-          {
-            capabilities: {
-              sampling: {},
-              roots: {
-                listChanged: true,
-              },
-            },
-          },
-        );
-        this.anthropic = new Anthropic({
-          apiKey: claudeApiKey,
-          dangerouslyAllowBrowser: true,
-        });
-      }
-      
-      async processQuery(query: string, tools: Tool[]): Promise<string> {
-        const messages: MessageParam[] = [
-          {
-            role: "user",
-            content: query,
-          },
-        ];
-      
-        const finalText: string[] = [];
-        const MAX_ITERATIONS = 5; 
-        let iteration = 0;
-      
-        // Helper function to recursively sanitize schema objects
-        const sanitizeSchema = (schema: unknown): unknown => {
-          if (!schema || typeof schema !== 'object') return schema;
-      
-          // Handle array
-          if (Array.isArray(schema)) {
-            return schema.map(item => sanitizeSchema(item));
-          }
-      
-          // Now we know it's an object
-          const schemaObj = schema as Record<string, unknown>;
-          const sanitized: Record<string, unknown> = {};
-      
-          for (const [key, value] of Object.entries(schemaObj)) {
-            if (key === 'properties' && value && typeof value === 'object' && !Array.isArray(value)) {
-              // Handle properties object
-              const propertiesObj = value as Record<string, unknown>;
-              const sanitizedProps: Record<string, unknown> = {};
-              const keyMapping: Record<string, string> = {};
-      
-              for (const [propKey, propValue] of Object.entries(propertiesObj)) {
-                const sanitizedKey = propKey.replace(/[^a-zA-Z0-9_-]/g, '_');
-                keyMapping[propKey] = sanitizedKey;
-                sanitizedProps[sanitizedKey] = sanitizeSchema(propValue);
-              }
-      
-              sanitized[key] = sanitizedProps;
-      
-              // Update required fields if they exist
-              if ('required' in schemaObj && Array.isArray(schemaObj.required)) {
-                sanitized.required = (schemaObj.required as string[]).map(
-                  (req: string) => keyMapping[req] || req
-                );
-              }
-            } else {
-              sanitized[key] = sanitizeSchema(value);
-            }
-          }
-      
-          return sanitized;
-        };
-      
-        const mappedTools = tools.map((tool: Tool) => {
-          // Deep copy and sanitize the schema
-          let inputSchema;
-          if (tool.input_schema) {
-            inputSchema = JSON.parse(JSON.stringify(tool.input_schema));
-          } else {
-            // If no input schema, create a basic object schema
-            inputSchema = {
-              type: "object",
-              properties: {},
-              required: []
-            };
-          }
-      
-          // Ensure the schema has a type field
-          if (!inputSchema.type) {
-            inputSchema.type = "object";
-          }
-      
-          // Ensure properties exists for object types
-          if (inputSchema.type === "object" && !inputSchema.properties) {
-            inputSchema.properties = {};
-          }
-      
-          const sanitizedSchema = sanitizeSchema(inputSchema);
-      
-          return {
-            name: tool.name,
-            description: tool.description,
-            input_schema: sanitizedSchema,
-          } as Tool;
-        });
-      
-        let response = await this.anthropic.messages.create({
-          model: "claude-3-haiku-20240307",
-          max_tokens: 1000,
-          messages,
-          tools: mappedTools,
-        });
-      
-        while (iteration < MAX_ITERATIONS) {
-          iteration++;
-          let hasToolUse = false;
-          
-          const assistantContent = [];
-          
-          for (const content of response.content) {
-            if (content.type === "text") {
-              finalText.push(content.text);
-              assistantContent.push(content);
-            } else if (content.type === "tool_use") {
-              hasToolUse = true;
-              assistantContent.push(content);
-              
-              try {
-                const toolName = content.name;
-                const toolArgs = content.input as { [x: string]: unknown } | undefined;
-      
-                const result = await this.callTool({
-                  name: toolName,
-                  arguments: toolArgs,
-                });
-                
-                finalText.push(
-                  `[Calling tool ${toolName} with args ${JSON.stringify(toolArgs)}]`
-                );
-      
-                if (assistantContent.length > 0) {
-                  messages.push({
-                    role: "assistant",
-                    content: assistantContent,
-                  });
-                }
-      
-                messages.push({
-                  role: "user",
-                  content: [
-                    {
-                      type: "tool_result",
-                      tool_use_id: content.id,
-                      content: result.content as string,
-                    }
-                  ],
-                });
-              } catch (error) {
-                console.error(`Tool ${content.name} failed:`, error);
-                finalText.push(`[Tool ${content.name} failed: ${error}]`);
-                
-                messages.push({
-                  role: "assistant",
-                  content: assistantContent,
-                });
-                
-                messages.push({
-                  role: "user",
-                  content: [
-                    {
-                      type: "tool_result",
-                      tool_use_id: content.id,
-                      content: `Error: ${error}`,
-                      is_error: true,
-                    }
-                  ],
-                });
-              }
-            }
-          }
-      
-          if (!hasToolUse) {
-            break;
-          }
-      
-          try {
-            response = await this.anthropic.messages.create({
-              model: "claude-3-5-sonnet-20241022",
-              max_tokens: 1000,
-              messages,
-              tools: mappedTools,
-            });
-          } catch (error) {
-            console.error("API call failed:", error);
-            finalText.push(`[API Error: ${error}]`);
-            break;
-          }
-        }
-      
-        for (const content of response.content) {
-          if (content.type === "text") {
-            finalText.push(content.text);
-          }
-        }
-      
-        if (iteration >= MAX_ITERATIONS) {
-          finalText.push(`[Warning: Reached maximum iterations (${MAX_ITERATIONS}). Stopping to prevent excessive API usage.]`);
-        }
-      
-        return finalText.join("\n");
-      }
-
-      async chatLoop(tools: Tool[]) {
-        const rl = readline.createInterface({
-          input: process.stdin,
-          output: process.stdout,
-        });
-      
-        try {
-          console.log("\nMCP Client Started!");
-          console.log("Type your queries or 'quit' to exit.");
-      
-          while (true) {
-            const message = await rl.question("\nQuery: ");
-            if (message.toLowerCase() === "quit") {
-              break;
-            }
-            const response = await this.processQuery(message, tools);
-            console.log("\n" + response);
-          }
-        } finally {
-          rl.close();
-        }
-      }
-      
-      async cleanup() {
-        await this.close();
-      }
-  
+  const connect = async (_e?: unknown, retryCount: number = 0, mcpClient?: MCPClient) => {
+    let client;
+    if (mcpClient){
+      client = mcpClient
     }
-    const client = new MCPClient();
+    else {
+      client = new MCPClient(claudeApiKey);
+    }
 
     try {
       await checkProxyHealth();
